Move migrations to /migrations folder
