+++ package.json
@@ -20,7 +20,7 @@
     "docs:api": "swagger-markdown -i docs/api.yml -o docs/api.md",
     "docs": "npm run docs:validate && npm run docs:api",
     "dev": "NODE_ENV=development nodemon --watch app app/index.js",
-    "migrate": "knex --knexfile app/db/knexfile.js migrate:latest",
+    "migrate": "knex migrate:latest",
     "test": "NODE_ENV=test nyc ava app/tests/**/*.test.js -c 1 --serial --verbose",
     "lint": "next lint",
     "build": "next build",
